requirements.txt
asgiref==3.6.0
Django==4.1.7
django-recaptcha==3.0.0
django-zxcvbn-password==2.1.1
sqlparse==0.4.3
tzdata==2023.2
zxcvbn==4.4.28
python-decouple-3.8